refactor(mappers): rename files to kebab-case
