+++ src/structured-data/handler.js
@@ -15,7 +15,7 @@ import { isArray } from '@adobe/spacecat-shared-utils';
 import { AuditBuilder } from '../common/audit-builder.js';
 import { syncSuggestions } from '../utils/data-access.js';
 import { convertToOpportunity } from '../common/opportunity.js';
-import { OpportunityData } from './opportunityDataMapper.js';
+import { OpportunityData } from './opportunity-data-mapper.js';
 
 const AUDIT_TYPE = 'structured-data';
