+++ src/structured-data/opportunity-data-mapper.js
